isolate ThemeArgs into a separate file
