+++ packages/zero-runtime/src/theme.d.ts
@@ -1,0 +1,1 @@
+export interface ThemeArgs {}
